requirements.txt
streamlit
openai
python-dotenv
pandas
scikit-learn
textblob
matplotlib
seaborn
